+++ CHANGELOG.rst
@@ -1,8 +1,10 @@
-
 Unreleased
 ----------
 - BREAKING CHANGE: dropping Python 3.6 support.
 - Fixed error in getting ``plain_text()`` of emptied-out wikitext (#113)
+- Deprecated: Calling ``Template.set_arg()`` without specifying a value  for ``preserve_spacing`` parameter is deprecated.
+  This is a temporary warning in preparation for changing the default value of this parameter from ``True`` to ``False``. (#111)
+- Fixed the ``stacklevel`` of warnings.
 
 v0.51.2
 -------